chore(plugin/antd): [Header] move to antd shared components
