+++ plugins/antd/src/components/layout/HeaderBarLeftSideMenuLayout/HeaderBarLeftSideMenuLayout.tsx
@@ -3,7 +3,7 @@ import { Layout } from 'antd';
 import dynamic from 'next/dynamic';
 import { FC } from 'react';
 import { useSelector } from 'react-redux';
-import { Header } from './Header';
+import { Header } from '../components';
 
 const SidebarMenu = dynamic(() => import('./SidebarMenu').then(mod => mod.SidebarMenu), {
   ssr: false,
