+++ plugins/antd/src/components/layout/components/index.ts
@@ -1,1 +1,2 @@
+export * from './Header';
 export * from './UserMenu';
